[refactor] SeatList 클래스 생성
