+++ api/src/main/java/org/flab/api/domain/event/domain/seat/SeatList.java
@@ -1,0 +1,16 @@
+package org.flab.api.domain.event.domain.seat;
+
+import java.util.List;
+
+public class SeatList {
+
+    private final List<Seat> seatList;
+
+    public SeatList(List<Seat> seatList) {
+        this.seatList = seatList;
+    }
+
+    public long countSeatByGradeId(long gradeId) {
+        return seatList.stream().filter(seat -> gradeId == seat.getGrade().getId()).count();
+    }
+}
